+++ domain-models-runtime/src/main/java/org/folio/rest/persist/PostgresClient.java
@@ -182,6 +182,7 @@ public class PostgresClient {
     } catch (Exception e) {
       log.error(e.getMessage(), e);
     }
+    postgresClient.vertx = vertx;
     return postgresClient;
   }