+++ domain-models-runtime/src/main/java/org/folio/rest/tools/client/HttpModuleClient2.java
@@ -11,6 +11,9 @@ import java.util.function.Function;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import org.folio.rest.persist.PostgresClient;
+import org.folio.rest.tools.client.exceptions.PopulateTemplateException;
+import org.folio.rest.tools.client.exceptions.PreviousRequestException;
 import org.folio.rest.tools.parser.JsonPathParser;
 import org.folio.rest.tools.utils.VertxUtils;
 
@@ -20,7 +23,9 @@ import com.google.common.cache.CacheStats;
 import com.google.common.cache.LoadingCache;
 
 import io.vertx.core.Handler;
+import io.vertx.core.MultiMap;
 import io.vertx.core.Vertx;
+import io.vertx.core.buffer.Buffer;
 import io.vertx.core.http.HttpClient;
 import io.vertx.core.http.HttpClientOptions;
 import io.vertx.core.http.HttpClientRequest;
@@ -37,6 +42,8 @@ import io.vertx.core.logging.LoggerFactory;
  */
 public class HttpModuleClient2 {
 
+  static LoadingCache<String, Response> cache = null;
+
   private static final String CTYPE = "Content-Type";
   private static final String ACCEPT = "Accept";
   private static final String APP_JSON_CTYPE = "application/json";
@@ -49,7 +56,6 @@ public class HttpModuleClient2 {
   private String tenantId;
   private HttpClientOptions options;
   private HttpClient httpClient;
-  private LoadingCache<String, CompletableFuture<Response>> cache = null;
   private Vertx vertx;
   private boolean autoCloseConnections = true;
   private Map<String, String> headers = new HashMap<>();
@@ -113,7 +119,7 @@ public class HttpModuleClient2 {
     headers.put(ACCEPT, APP_JSON_ACCEPT);
   }
 
-  private void request(HttpMethod method, String endpoint, Map<String, String> headers,
+  private void request(HttpMethod method, Buffer data, String endpoint, Map<String, String> headers,
       boolean cache, Handler<HttpClientResponse> responseHandler, CompletableFuture<Response> cf2){
 
     try {
@@ -133,7 +139,12 @@ public class HttpModuleClient2 {
         this.headers.putAll(headers);
       }
       request.headers().setAll(this.headers);
-      request.end();
+      if(data != null){
+        request.end(data);
+      }
+      else{
+        request.end();
+      }
     } catch (Exception e) {
       Response r = new Response();
       r.populateError(endpoint, -1, e.getMessage());
@@ -142,7 +153,7 @@ public class HttpModuleClient2 {
     }
   }
 
-  public CompletableFuture<Response> request(HttpMethod method, String endpoint, Map<String, String> headers, RollBackURL rollbackURL,
+  public CompletableFuture<Response> request(HttpMethod method, Buffer data, String endpoint, Map<String, String> headers, RollBackURL rollbackURL,
       boolean cachable, BuildCQL bCql) throws Exception {
 
     if(bCql != null){
@@ -150,16 +161,16 @@ public class HttpModuleClient2 {
     }
     if(cachable){
       initCache();
-      CompletableFuture<Response> j = cache.get(endpoint);
-      if(j != null && !j.isCompletedExceptionally() && j.isDone()){
-        return j;
+      Response j = cache.get(endpoint);
+      if(j != null && j.getBody() != null){
+        CompletableFuture<Response> cf = new CompletableFuture<>();
+        cf.complete(j);
+        log.debug("entry retrieved from cache: " + endpoint);
+        return cf;
       }
     }
     CompletableFuture<Response> cf = new CompletableFuture<>();
     HTTPJsonResponseHandler handler = new HTTPJsonResponseHandler(endpoint, cf);
-    if(cachable) {
-      handler.cache = cache;
-    }
     if(autoCloseConnections){
       handler.httpClient = httpClient;
     }
@@ -167,9 +178,30 @@ public class HttpModuleClient2 {
       handler.rollbackURL = rollbackURL;
     }
 
-    request(method, endpoint, headers, cachable, handler, cf);
+    request(method, data, endpoint, headers, cachable, handler, cf);
 
     return cf;
+  }
+
+  public CompletableFuture<Response> request(HttpMethod method, String endpoint, Map<String, String> headers, RollBackURL rollbackURL,
+      boolean cachable, BuildCQL bCql) throws Exception {
+
+      return request(method, null, endpoint, headers, rollbackURL,cachable, bCql);
+  }
+
+  public CompletableFuture<Response> request(HttpMethod method, Buffer data, String endpoint, Map<String, String> headers)
+      throws Exception {
+    return request(method, data, endpoint, headers, null, false, null);
+  }
+
+  public CompletableFuture<Response> request(HttpMethod method, Object pojo, String endpoint, Map<String, String> headers)
+      throws Exception {
+    return request(method, Buffer.buffer(PostgresClient.pojo2json(pojo)), endpoint, headers, null, false, null);
+  }
+
+  public CompletableFuture<Response> request(HttpMethod method, String endpoint, Map<String, String> headers)
+      throws Exception {
+    return request(method, endpoint, headers, null, false, null);
   }
 
   public CompletableFuture<Response> request(String endpoint, Map<String, String> headers, boolean cache, BuildCQL cql)
@@ -212,14 +244,48 @@ public class HttpModuleClient2 {
     return request(HttpMethod.GET, endpoint, null, null, true, null);
   }
 
+  /**
+   * A request that should be used within a thenCompose() completable future call and receive as
+   * input a Response object from the previous request.
+   * The chainedRequest will
+   * 1. callback to the passed in Consumer with the Response of the previous request for handling
+   * 2. if the passed in Response contains errors - the current request will not be sent and a
+   * completable future will be returned indicating that the request was not run
+   * 3. replace placeholder in the url {a.b[0]} with actual values appearing in the json passed in
+   * by the thenCompose(). For example: passing in:
+   * http://localhost:9130/users/{users[0].username}
+   * will look in the passed in json for the value found in the first user in a json array[].username
+   * NOTE that if a template has a placeholder, and the path indicated in the placeholder does not
+   * exist in the passed in Response - the current request will not be sent and a completeable future
+   * with a Response (containing the error) will be returned
+   * 4. build a cql via values in the passed in json
+   * 5. Send the request
+   * @param urlTempate
+   * @param headers
+   * @param inheritOkapiHeaders - take all okapi headers in passed in Response and over write / add to this request
+   * @param cql
+   * @param processPassedInResponse
+   * @return
+   */
   public Function<Response, CompletableFuture<Response>> chainedRequest(
-      String urlTempate, Map<String, String> headers, BuildCQL cql, Consumer<Response> processPassedInResponse){
+      String urlTempate, Map<String, String> headers, boolean inheritOkapiHeaders, BuildCQL cql, Consumer<Response> processPassedInResponse){
     try {
       List<String> replace = getTagValues(urlTempate);
       return (resp) -> {
         //once future completes we enter this section//
         //create a new request based on the content of the passed in response (resp)//
         try {
+          //call back to the passed in consumer, this function should analyze the returned//
+          //response for errors / exceptions and return accordingly if found//
+          if(processPassedInResponse != null){
+            processPassedInResponse.accept(resp);
+          }
+          if(resp.getError() != null || resp.getException() != null){
+            CompletableFuture<Response> cf = new CompletableFuture<>();
+            PreviousRequestException pre = new PreviousRequestException("for template, " + urlTempate);
+            cf.complete(createResponse(urlTempate, pre));
+            return cf;
+          }
           int size = replace.size();
           String newURL = urlTempate;
           if(size > 0){
@@ -228,37 +294,61 @@ public class HttpModuleClient2 {
               String val = (String)jpp.getValueAt(replace.get(i));
               if(val == null){
                 log.error("Unable to replace {"+replace.get(i)+"} with content from received json result, does the json contain this field? Endpoint: "+ resp.endpoint);
-                throw new Exception("Missing {"+replace.get(i)+"} value from results received from endpoint " + resp.endpoint);
+                PopulateTemplateException pe = new PopulateTemplateException("Missing {"+replace.get(i)+"} value from results received from endpoint " + resp.endpoint);
+                CompletableFuture<Response> cf = new CompletableFuture<>();
+                cf.complete(createResponse(urlTempate, pe));
+                return cf;
               }
               newURL = urlTempate.replace("{"+replace.get(i)+"}", val);
             }
           }
-          //call back to the passed in consumer, this function should analyze the returned//
-          //response for errors / exceptions and return accordingly if found//
-          processPassedInResponse.accept(resp);
-          if(resp.getError() != null || resp.getException() != null){
-            return null;
+          if(cql != null){
+            cql.setResponse(resp);
           }
-          else{
-            if(cql != null){
-              cql.setResponse(resp);
-            }
-            //call request//
-            return request(newURL, headers, cql);
+          //call request//
+          if(inheritOkapiHeaders){
+            mergeHeaders(headers, resp.headers, resp.endpoint);
           }
+          return request(newURL, headers, cql);
         } catch (Exception e) {
-          resp.exception = e;
-          resp.endpoint = urlTempate;
-          resp.body = null;
-          resp.code = -1;
-          processPassedInResponse.accept(resp);
+          CompletableFuture<Response> cf = new CompletableFuture<>();
+          cf.complete(createResponse(urlTempate, e));
+          return cf;
         }
-        return null;
       };
     } catch (Exception e) {
       e.printStackTrace();
     }
     return null;
+  }
+
+  private void mergeHeaders(Map<String, String >headers1,  MultiMap headers2, String endpoint){
+    Consumer<Map.Entry<String,String>> consumer = entry -> {
+      String headerKey = entry.getKey().toLowerCase();
+      if(headerKey.startsWith("x-okapi")){
+        headers1.put(headerKey, entry.getValue());
+      }
+    };
+    if(headers != null){
+      headers2.forEach(consumer);
+    }
+    else{
+      log.warn("headers passed into chainedRequest are null, from endpoint " + endpoint);
+    }
+  }
+
+  public Function<Response, CompletableFuture<Response>> chainedRequest(
+      String urlTempate, Map<String, String> headers, BuildCQL cql, Consumer<Response> processPassedInResponse){
+    return chainedRequest(urlTempate, headers, false, cql, processPassedInResponse);
+  }
+
+  private Response createResponse(String urlTempate, Exception e){
+    Response r = new Response();
+    r.exception = e;
+    r.endpoint = urlTempate;
+    r.body = null;
+    r.code = -1;
+    return r;
   }
 
   private static List<String> getTagValues(final String str) {
@@ -295,10 +385,10 @@ public class HttpModuleClient2 {
         .maximumSize(1000)
         .expireAfterWrite(cacheTO, TimeUnit.MINUTES)
         .build(
-          new CacheLoader<String, CompletableFuture<Response>>() {
+          new CacheLoader<String, Response>() {
             @Override
-            public CompletableFuture<Response> load(String key) throws Exception {
-              return new CompletableFuture<>();
+            public Response load(String key) throws Exception {
+              return new Response();
             }
           });
     }