+++ domain-models-runtime/src/main/java/org/folio/rest/tools/client/exceptions/PopulateTemplateException.java
@@ -1,0 +1,18 @@
+package org.folio.rest.tools.client.exceptions;
+
+/**
+ * @author shale
+ *
+ */
+public class PopulateTemplateException extends Exception {
+
+  private static final long serialVersionUID = 4773848346592449712L;
+
+  /**
+   * @param message
+   */
+  public PopulateTemplateException(String message) {
+    super(message);
+  }
+
+}