+++ domain-models-runtime/src/main/java/org/folio/rest/persist/PostgresClient.java
@@ -17,6 +17,8 @@ import java.util.regex.Pattern;
 
 import javax.crypto.SecretKey;
 
+import org.apache.commons.collections4.map.HashedMap;
+import org.apache.commons.collections4.map.MultiKeyMap;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.text.StrSubstitutor;
@@ -93,7 +95,7 @@ public class PostgresClient {
   private static boolean         embeddedMode             = false;
   private static String          configPath               = null;
   private static ObjectMapper    mapper                   = ObjectMapperTool.getMapper();
-  private static Map<String, PostgresClient> connectionPool = new HashMap<>();
+  private static MultiKeyMap<Object, PostgresClient> connectionPool = MultiKeyMap.multiKeyMap(new HashedMap<>());
   private static String moduleName                        = null;
 
   private static final String CLOSE_FUNCTION_POSTGRES = "WINDOW|IMMUTABLE|STABLE|VOLATILE|"
@@ -169,11 +171,11 @@ public class PostgresClient {
    */
   private static PostgresClient getInstanceInternal(Vertx vertx, String tenantId) {
     // assumes a single thread vertx model so no sync needed
-    PostgresClient postgresClient = connectionPool.get(tenantId);
+    PostgresClient postgresClient = connectionPool.get(vertx, tenantId);
     try {
       if (postgresClient == null) {
         postgresClient = new PostgresClient(vertx, tenantId);
-        connectionPool.put(tenantId, postgresClient);
+        connectionPool.put(vertx, tenantId, postgresClient);
       }
       if (postgresClient.client == null) {
         // in connectionPool, but closeClient() has been invoked
@@ -182,7 +184,6 @@ public class PostgresClient {
     } catch (Exception e) {
       log.error(e.getMessage(), e);
     }
-    postgresClient.vertx = vertx;
     return postgresClient;
   }
 
@@ -259,7 +260,7 @@ public class PostgresClient {
     }
     AsyncSQLClient clientToClose = client;
     client = null;
-    connectionPool.remove(tenantId);  // remove (tenantId, this) entry
+    connectionPool.removeMultiKey(vertx, tenantId);  // remove (vertx, tenantId, this) entry
     clientToClose.close(whenDone);
   }
 
@@ -1609,24 +1610,13 @@ public class PostgresClient {
   }
 
   /**
-   *
    * Will connect to a specific database and execute the commands in the .sql file
-   * against that database  -
-   *
+   * against that database.<p />
    * NOTE: NOT tested on all types of statements - but on a lot
    *
-   * @param sqlFile - reader to sql file with executable statements
-   * @param newDB - if creating a new database is included in the file - include the name of the db as after running the
-   * create database command appearing in the file, there will be a new connection created to the
-   * newDB and all subsequent commands will be executed against the newly created newDB name. the user / password of
-   * the connection is currently hard coded to the value of newDB - so the .sql file which is creating the DB
-   * should do something like this:
-        CREATE DATABASE myuniversity
-            WITH OWNER = myuniversity
-            PASSWORD = myuniversity
+   * @param sqlFile - string of sqls with executable statements
    * @param stopOnError - stop on first error
-   * @param timeout - in seconds
-   * @param replyHandler - list of statements that failed - if any
+   * @param replyHandler - the handler's result is the list of statements that failed; the list may be empty
    */
   public void runSQLFile(String sqlFile, boolean stopOnError,
       Handler<AsyncResult<List<String>>> replyHandler){
@@ -1641,7 +1631,6 @@ public class PostgresClient {
       List<String> execStatements = new ArrayList<>();
       boolean inFunction = false;
       boolean inCopy = false;
-      boolean funcCompleteAddFuncAttributes = false;
       for (int i = 0; i < allLines.length; i++) {
         if(allLines[i].toUpperCase().matches("^\\s*(CREATE USER|CREATE ROLE).*") && AES.getSecretKey() != null) {
           final Pattern pattern = Pattern.compile("PASSWORD\\s*'(.+?)'\\s*", Pattern.CASE_INSENSITIVE);
@@ -1699,6 +1688,10 @@ public class PostgresClient {
           }
           singleStatement.append(" " + allLines[i]);
         }
+      }
+      String lastStatement = singleStatement.toString();
+      if (! lastStatement.trim().isEmpty()) {
+        execStatements.add(lastStatement);
       }
       execute(execStatements.toArray(new String[]{}), stopOnError, replyHandler);
     } catch (Exception e) {